+++ src/vis/vishooks/math/index.ts
@@ -1,4 +1,4 @@
-import * as m4 from './m4';
+import * as m4 from './matrix4x4';
 import * as v3 from './v3';
 
 export { m4, v3 };
