+++ src/vis/vishooks/math/vectorv3.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-param-reassign */
 import { Vector3 } from './interfaces';
 
-export function identityMatrixV3(out = new Float32Array(3)): Vector3 {
+export function vector3(out = new Float32Array(3)): Vector3 {
   out[0] = 0;
   out[1] = 0;
   out[2] = 0;
